BlitBuffer: Reformat in a C99 fashion
* Tighter scope, const, restricted pointers
* Avoid array indexing where possible
* Use unsigned data types for dimensions/sizes
  That's really what the code expects, and everything is already
  bounds-checked on the Lua side.
* Resync dither_o8x8 w/ FBInk
* In the few places that weren't already doing so,
  cadge a few shifts by relying on pointer arithmetic rules,
  much like BB_GET_PIXEL
* Implement ditheralphablitFrom (straight alpha)
* Move the helper macros to the header, it's useful for C consumers

* ffi/blitbuffer: Clear up a c/p snafu w/ invertBlitFrom
* ffi/blitbuffer: Good-bye, lovingly handcrafted BMP encoding ;)

+++ djvu.c
@@ -676,7 +676,7 @@ static int drawPage(lua_State *L) {
 		}
 	}
 	ddjvu_format_set_gamma(page->doc->pixelformat, gamma);
-	int bbsize = (bb->w)*(bb->h)*page->doc->pixelsize;
+	size_t bbsize = (bb->w)*(bb->h)*page->doc->pixelsize;
 	uint8_t *imagebuffer = bb->data;
 
 	/*printf("@page %d, @@zoom:%f, offset: (%d, %d)\n", page->num, dc->zoom, dc->offset_x, dc->offset_y);*/
